requirements.txt
numpy
scikit-learn
sentence-transformers
pypdf
tqdm
python-dotenv
rich
google-genai
pytest